+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/XmlCompleteConfigurationTest.php
@@ -15,6 +15,9 @@ use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 
+/**
+ * @group legacy
+ */
 class XmlCompleteConfigurationTest extends CompleteConfigurationTestCase
 {
     public function testFirewallPatterns()