+++ src/Symfony/Component/Routing/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Allow aliases and deprecations in `#[Route]` attribute
  * Add the `Requirement::MONGODB_ID` constant to validate MongoDB ObjectIDs in hexadecimal format
+ * Deprecate `XmlFileLoader`
 
 7.2
 ---